Added routing and a couple sample pages for the profile creation pages, modeled after home.dart
Now if you want to add a page to the profile creation part, add your dart file under the profileCreation directory. Then go into ProfileCreationParent, and add as a tab the label and the widget you want. Then down in the BottomNavigationBar section you can add an icon to represent that link.